# Don't use `deepcopy_internal(xi, stackdict)` for bits fields