--- a/base/deepcopy.jl
+++ b/base/deepcopy.jl
@@ -65,7 +65,7 @@ function deepcopy_internal(@nospecialize(x), stackdict::IdDict)
         for i in 1:nf
             if isdefined(x, i)
                 xi = getfield(x, i)
-                xi = deepcopy_internal(xi, stackdict)::typeof(xi)
+                !isbits(xi) && (xi = deepcopy_internal(xi, stackdict)::typeof(xi))
                 ccall(:jl_set_nth_field, Cvoid, (Any, Csize_t, Any), y, i-1, xi)
             end
         end
@@ -76,7 +76,7 @@ function deepcopy_internal(@nospecialize(x), stackdict::IdDict)
         for i in 1:nf
             if isdefined(x, i)
                 xi = getfield(x, i)
-                xi = deepcopy_internal(xi, stackdict)::typeof(xi)
+                !isbits(xi) && (xi = deepcopy_internal(xi, stackdict)::typeof(xi))
                 flds[i] = xi
             else
                 nf = i - 1 # rest of tail must be undefined values