@Tortar 2023-12-21
Removing the nospecialize macro to the x argument of deepcopy_internal also works (actually with more nested object it improves even more the performance) but there is surely a good reason why it is there, I guess it is related to the nesting itself