@vtjnash 2023-12-21
Looks like we already did this for Array/Memory, so good to be more consistent here